agency-scout/requirements.txt
fastapi
uvicorn
pandas
requests
beautifulsoup4
builtwith
python-whois
jinja2
python-multipart
googlemaps
